+++ src/apis/products/products.ts
@@ -1,9 +1,8 @@
 import axiosInstance from '@/apis/axios';
-import type { ProductDetailResponse, RankingProductsResponse } from '@/apis/products/type';
 import type { RankingFilterOption } from '@/types';
 
 export const getRankingProducts = async (filterOption: RankingFilterOption) => {
-  const response = await axiosInstance.get<RankingProductsResponse>('/api/v1/ranking/products', {
+  const response = await axiosInstance.get<Home.RankingProductsResponse>('/api/v1/ranking/products', {
     params: {
       targetType: filterOption.targetType || 'ALL',
       rankType: filterOption.rankType || 'MANY_WISH_RECEIVE',
@@ -13,7 +12,7 @@ export const getRankingProducts = async (filterOption: RankingFilterOption) => {
 };
 
 export const getProductDetail = async (productId: string) => {
-  const response = await axiosInstance.get<ProductDetailResponse>(
+  const response = await axiosInstance.get<Home.ProductDetailResponse>(
     `/api/v1/products/${productId}/detail`,
   );
   return response.data;
