+++ src/apis/themes/themes.ts
@@ -1,9 +1,7 @@
 import axiosInstance from '@/apis/axios';
-import type { ThemeProductsResponse, ThemesResponse } from '@/apis/themes/type';
-import type { GetThemeProductsType } from '@/apis/themes/type';
 
 export const getThemes = async () => {
-  const response = await axiosInstance.get<ThemesResponse>('/api/v1/themes');
+  const response = await axiosInstance.get<Theme.ThemesResponse>('/api/v1/themes');
   return response.data;
 };
 
@@ -11,8 +9,8 @@ export const getThemeProducts = async ({
   themeKey,
   pageToken,
   maxResults = 10,
-}: GetThemeProductsType) => {
-  const response = await axiosInstance.get<ThemeProductsResponse>(
+}: Theme.GetThemeProductsType) => {
+  const response = await axiosInstance.get<Theme.ThemeProductsResponse>(
     `/api/v1/themes/${themeKey}/products`,
     {
       params: {
