+++ src/pages/Login/login.d.ts
@@ -1,0 +1,3 @@
+declare namespace Login {
+
+}
