+++ src/types/index.ts
@@ -1,12 +1,4 @@
-export type ThemeData = {
-  id: number;
-  key: string;
-  label: string;
-  title: string;
-  description?: string;
-  backgroundColor: string;
-  imageURL?: string;
-};
+
 
 export const validTargetTypes = ['ALL', 'FEMALE', 'MALE', 'TEEN'] as const;
 export const validRankTypes = ['MANY_WISH', 'MANY_RECEIVE', 'MANY_WISH_RECEIVE'] as const;
